Add run-tests-android.sh script. Fix tests on android.
Add a new script run-tests-android.sh. It runs tests through ADB and
handles all calls to the C compiler (through run-compiler.sh or make).
It parses the tests' output to see if make or the C compiler needs to
be called, so run-compiler.sh's output must go to stdout.

Disable failing / inapplicable tests:
- elfcore.test.sh, elf-sans-immobile.test.sh
  Android only supports position-independent executables.
- :INTERRUPT, :OVEREAGER-CHARACTER-BUFFERING from stream.impure.lisp and fifo-slow.impure.lisp
  On unrooted devices MKFIFO signals an error "Permission denied".