Workaround for mprotect EACCES. FIXME: find out why this only happens on 32-bit?
On Android it is disallowed to add executable permissions for a dirty file page.
We simply get rid of all file pages here by reading the core instead of mmaping.

For some unknown reason some of the mmaped exec core pages in the gc space
during the warm init are modified, then protected with PROT_READ, and then
attempted to be unprotected during a wp page fault. This mysteriously only
happens on 32-bit android.